stt-service/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-speech
pydub
python-multipart
